fix(deposit): return main component and move info box
